fixed printf to work with more integer types

+++ src/kernel.cpp
@@ -44,7 +44,7 @@ extern "C" void main() {
 
     for (uint16_t i = 0; i < 6; i++) {
         for (uint16_t j = 0; j < 6; j++) {
-            uint16_t word = pciConfigRead32(i, j, 0, 0);
+            uint32_t word = pciConfigRead32(i, j, 0, 0);
             // fprintf(Serial, "word is: 0x%x %x\n", word<<16, word & 0xffff);
             fprintf(Serial, "word is: 0x%x \n", word);
         }
